+++ app/src/main/java/online/taxcore/pos/data/params/InvoiceRequest.kt
@@ -22,8 +22,3 @@ class PaymentItem(
     var amount: Double = 0.0,
     var paymentType: String = ""
 )
-
-enum class InvoiceRequestType {
-    VSDC,
-    ESDC
-}