+++ app/src/main/java/online/taxcore/pos/ui/catalog/ItemDetailActivity.kt
@@ -17,8 +17,19 @@ import com.pawegio.kandroid.longToast
 import com.vicpin.krealmextensions.queryAndUpdate
 import com.vicpin.krealmextensions.queryFirst
 import com.vicpin.krealmextensions.save
-import kotlinx.android.synthetic.main.app_bar_main.*
-import kotlinx.android.synthetic.main.item_details_activity.*
+import kotlinx.android.synthetic.main.app_bar_main.toolbar
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_list_vat
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_barcode
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_input_barcode
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_input_name
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_input_price
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_name
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_price
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_scroll
+import kotlinx.android.synthetic.main.item_details_activity.activity_detail_plu_vat
+import kotlinx.android.synthetic.main.item_details_activity.invalid_taxes_label
+import kotlinx.android.synthetic.main.item_details_activity.invalid_taxes_list
+import kotlinx.android.synthetic.main.item_details_activity.ll_invalid_taxes
 import online.taxcore.pos.R
 import online.taxcore.pos.data.PrefService
 import online.taxcore.pos.data.local.TaxesManager
@@ -34,7 +45,7 @@ import online.taxcore.pos.ui.common.TaxesCheckedAdapter
 import online.taxcore.pos.ui.invoice.InvoiceFragment
 import online.taxcore.pos.ui.invoice.InvoiceFragment.Companion.BARCODE_EAN_EXTRA
 import online.taxcore.pos.utils.hideKeyboard
-import org.jetbrains.anko.contentView
+
 import javax.inject.Inject
 
 class ItemDetailActivity : BaseActivity() {
@@ -60,8 +71,6 @@ class ItemDetailActivity : BaseActivity() {
 
     private var updatedFieldsMap = mutableMapOf<String, Boolean>()
     private var useESDC: Boolean = false
-
-    private var existInvalidTax = false
 
     companion object {
         private const val EXTRA_ITEM_UUID = "EXTRA_ITEM_UUID"
@@ -116,6 +125,7 @@ class ItemDetailActivity : BaseActivity() {
             finish()
             true
         }
+
         R.id.actionItemSave -> {
             if (isInCreateMode and isFormValid()) {
                 saveItem()
@@ -125,6 +135,7 @@ class ItemDetailActivity : BaseActivity() {
 
             true
         }
+
         else -> super.onOptionsItemSelected(item)
     }
 
@@ -134,7 +145,8 @@ class ItemDetailActivity : BaseActivity() {
         supportActionBar?.setDisplayShowHomeEnabled(true)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
         supportActionBar?.setHomeButtonEnabled(true)
-        supportActionBar?.title = if (isInCreateMode) getString(R.string.add_item) else getString(R.string.edit_item)
+        supportActionBar?.title =
+            if (isInCreateMode) getString(R.string.add_item) else getString(R.string.edit_item)
     }
 
     private fun initFields() {
@@ -197,16 +209,16 @@ class ItemDetailActivity : BaseActivity() {
         val itemTaxLabels = item?.tax?.map { it.code } ?: arrayListOf()
 
         val appliedTaxItems = appTaxList
-                .map {
-                    if (itemTaxLabels.contains(it.code)) {
-                        it.isChecked = true
-                    }
-                    it
+            .map {
+                if (itemTaxLabels.contains(it.code)) {
+                    it.isChecked = true
                 }
+                it
+            }
 
         val appTaxLabels = appTaxList.map { it.code }
         val invalidTaxItems = item?.tax?.filter { appTaxLabels.contains(it.code).not() }
-                ?: arrayListOf()
+            ?: arrayListOf()
 
         if (invalidTaxItems.isNotEmpty()) {
             invalid_taxes_label.visibility = View.VISIBLE
@@ -257,7 +269,8 @@ class ItemDetailActivity : BaseActivity() {
             if (isBarcodeValid) {
                 activity_detail_plu_barcode.isErrorEnabled = false
             } else {
-                activity_detail_plu_barcode.error = getString(R.string.error_minimum_eight_characters)
+                activity_detail_plu_barcode.error =
+                    getString(R.string.error_minimum_eight_characters)
             }
 
             validateInputForm()
@@ -265,11 +278,14 @@ class ItemDetailActivity : BaseActivity() {
     }
 
     private fun setOnPriceChangeHandler() {
-        activity_detail_plu_input_price.filters = arrayOf<InputFilter>(DecimalDigitsInputFilter(12, 2))
+        activity_detail_plu_input_price.filters =
+            arrayOf<InputFilter>(DecimalDigitsInputFilter(12, 2))
         activity_detail_plu_input_price.onTextChanged {
 
             if (isNumber(it)) {
-                updatedFieldsMap["price"] = (it.roundTo2DecimalPlaces().toDouble() != item?.price?.roundToDecimal()?.toDouble())
+                updatedFieldsMap["price"] =
+                    (it.roundTo2DecimalPlaces().toDouble() != item?.price?.roundToDecimal()
+                        ?.toDouble())
             }
 
             isPriceValid = it.isNotEmpty() && isNumber(it)
@@ -333,7 +349,9 @@ class ItemDetailActivity : BaseActivity() {
         }
     }
 
-    private fun isFormValid() = hasTaxLabelApplied() && isPriceValid && isNameValid && isBarcodeValid
+    private fun isFormValid() =
+        hasTaxLabelApplied() && isPriceValid && isNameValid && isBarcodeValid
+
     private fun isFormUpdated(): Boolean = updatedFieldsMap.any { item -> item.value }
 
     private fun hasTaxLabelApplied(): Boolean {
@@ -343,10 +361,12 @@ class ItemDetailActivity : BaseActivity() {
 
         if (isInEditMode) {
             val appliedLabelsList = appliedTaxes?.toList()?.map { it.code } as Collection<String>
-            val appliedInvalidLabelsList = appliedInvalidTaxes?.toList()?.map { it.code } as Collection<String>
+            val appliedInvalidLabelsList =
+                appliedInvalidTaxes?.toList()?.map { it.code } as Collection<String>
             val itemLabelsList = item?.tax?.toList()?.map { it.code } as Collection<String>
 
-            updatedFieldsMap["appliedTaxes"] = appliedLabelsList != itemLabelsList || appliedInvalidLabelsList != itemLabelsList
+            updatedFieldsMap["appliedTaxes"] =
+                appliedLabelsList != itemLabelsList || appliedInvalidLabelsList != itemLabelsList
         }
 
         return appliedTaxes.isNullOrEmpty().not() || appliedInvalidTaxes.isNullOrEmpty().not()
@@ -432,9 +452,10 @@ class ItemDetailActivity : BaseActivity() {
 
         hideKeyboard()
 
-        contentView?.let {
+
+        activity_detail_plu_scroll?.let {
             Snackbar.make(it, R.string.toast_item_updated, Snackbar.LENGTH_SHORT)
-                    .show()
+                .show()
         }
     }