+++ app/src/main/java/online/taxcore/pos/utils/CtxUtils.kt
@@ -16,13 +16,9 @@ class CtxUtils(base: Context) : ContextWrapper(base) {
             var context = c
             val resources: Resources = context.resources
             val configuration: Configuration = resources.configuration
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                val localeList = LocaleList(localeToSwitchTo)
-                LocaleList.setDefault(localeList)
-                configuration.setLocales(localeList)
-            } else {
-                configuration.locale = localeToSwitchTo
-            }
+            val localeList = LocaleList(localeToSwitchTo)
+            LocaleList.setDefault(localeList)
+            configuration.setLocales(localeList)
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
                 context = context.createConfigurationContext(configuration)
             } else {